FILE: moves app-routes file into app/ folder
