+++ apps/spirit-islander/src/main.ts
@@ -4,7 +4,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { provideRouter } from '@angular/router';
 
 import { APP_NAME_TOKEN } from '@atocha/core/data-access';
-import { APP_ROUTES } from './app-routes';
+import { APP_ROUTES } from './app/app-routes';
 import { AppComponent } from './app/app.component';
 import { environment } from './environments/environment';
